Fix redundant package names
